requirements.txt
playwright
beautifulsoup4
python-dotenv
lxml
flask
pgeocode
